\name{GSEA-package}
\alias{GSEA-package}
\alias{GSEA}
\docType{package}
\title{
\packageTitle{GSEA}
}
\description{
\packageDescription{GSEA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GSEA}
\packageIndices{GSEA}
}
\author{
\packageAuthor{GSEA}

Maintainer: \packageMaintainer{GSEA}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{GSEA}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
