% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.Analyze.Sets.R
\name{GSEA.Analyze.Sets}
\alias{GSEA.Analyze.Sets}
\title{Performs leading edge analysis of a GSEA result}
\usage{
GSEA.Analyze.Sets(directory = getwd(), topgs = 20,
  non.interactive.run = F, height = 16, width = 16,
  gsea.type = "GSEA", doc.string = "gsea_result")
}
\arguments{
\item{directory}{the directory containing GSEA result files}

\item{gsea.type}{the gsea run type (either "GSEA" or "preranked")}

\item{doc.string}{the file prefixed in the original analysis}
}
\description{
`GSEA.Analyze.Sets` returns leading edge plots and gcts extracted from a GSEA result folder
}
\details{
This function is designed to be invoked immediately after `GSEA` to run leading edge analysis.
}
