% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.CollapseDataset.R
\name{GSEA.CollapseDataset}
\alias{GSEA.CollapseDataset}
\title{Maps user supplied identifiers to Gene Symbols}
\usage{
GSEA.CollapseDataset(dataplatform = chip, gct = dataset,
  collapse.mode = collapsemode)
}
\description{
`GSEA.CollapseDataset` applies a .CHIP file to the dataset to gene symbols using a specified collapse function
}
\details{
Internal `GSEA` function invoked if collapse.dataset == TRUE.
}
